01_etl/requirements.txt
pydantic==1.9.0
python-dotenv==0.19.2
psycopg2==2.9.3
elasticsearch>=7.0.0,<8.0.0